+++ tests/integration-tests/tests/name_server_pool_tests.rs
@@ -1,3 +1,4 @@
+extern crate futures;
 extern crate tokio;
 extern crate trust_dns;
 extern crate trust_dns_integration;
@@ -6,7 +7,12 @@ extern crate trust_dns_resolver;
 
 use std::net::*;
 use std::str::FromStr;
+use std::sync::{
+    atomic::{AtomicIsize, Ordering},
+    Arc,
+};
 
+use futures::future::{self, Future, Loop};
 use tokio::runtime::current_thread::Runtime;
 
 use trust_dns::op::Query;
@@ -18,25 +24,47 @@ use trust_dns_resolver::config::*;
 use trust_dns_resolver::name_server_pool::{ConnectionProvider, NameServer, NameServerPool};
 
 #[derive(Clone)]
-struct MockConnProvider {}
+struct MockConnProvider<O: OnSend> {
+    on_send: O,
+}
 
-impl ConnectionProvider for MockConnProvider {
-    type ConnHandle = MockClientHandle;
-
-    fn new_connection(_: &NameServerConfig, _: &ResolverOpts) -> Self::ConnHandle {
-        MockClientHandle::mock(vec![])
+impl Default for MockConnProvider<DefaultOnSend> {
+    fn default() -> Self {
+        Self {
+            on_send: DefaultOnSend,
+        }
     }
 }
 
-type MockedNameServer = NameServer<MockClientHandle, MockConnProvider>;
-type MockedNameServerPool = NameServerPool<MockClientHandle, MockConnProvider>;
+impl<O: OnSend> ConnectionProvider for MockConnProvider<O> {
+    type ConnHandle = MockClientHandle<O>;
+
+    fn new_connection(&self, _: &NameServerConfig, _: &ResolverOpts) -> Self::ConnHandle {
+        MockClientHandle::mock_on_send(vec![], self.on_send.clone())
+    }
+}
+
+type MockedNameServer<O> = NameServer<MockClientHandle<O>, MockConnProvider<O>>;
+type MockedNameServerPool<O> = NameServerPool<MockClientHandle<O>, MockConnProvider<O>>;
 
 #[cfg(test)]
 fn mock_nameserver(
     messages: Vec<ProtoResult<DnsResponse>>,
     options: ResolverOpts,
-) -> MockedNameServer {
-    let client = MockClientHandle::mock(messages);
+) -> MockedNameServer<DefaultOnSend> {
+    mock_nameserver_on_send(messages, options, DefaultOnSend)
+}
+
+#[cfg(test)]
+fn mock_nameserver_on_send<O: OnSend>(
+    messages: Vec<ProtoResult<DnsResponse>>,
+    options: ResolverOpts,
+    on_send: O,
+) -> MockedNameServer<O> {
+    let conn_provider = MockConnProvider {
+        on_send: on_send.clone(),
+    };
+    let client = MockClientHandle::mock_on_send(messages, on_send);
 
     NameServer::from_conn(
         NameServerConfig {
@@ -46,25 +74,42 @@ fn mock_nameserver(
         },
         options,
         client,
+        conn_provider,
     )
 }
 
 #[cfg(test)]
 fn mock_nameserver_pool(
-    udp: Vec<MockedNameServer>,
-    tcp: Vec<MockedNameServer>,
-    _mdns: Option<MockedNameServer>,
+    udp: Vec<MockedNameServer<DefaultOnSend>>,
+    tcp: Vec<MockedNameServer<DefaultOnSend>>,
+    _mdns: Option<MockedNameServer<DefaultOnSend>>,
     options: ResolverOpts,
-) -> MockedNameServerPool {
+) -> MockedNameServerPool<DefaultOnSend> {
+    mock_nameserver_pool_on_send::<DefaultOnSend>(udp, tcp, _mdns, options, DefaultOnSend)
+}
+
+#[cfg(test)]
+fn mock_nameserver_pool_on_send<O: OnSend>(
+    udp: Vec<MockedNameServer<O>>,
+    tcp: Vec<MockedNameServer<O>>,
+    _mdns: Option<MockedNameServer<O>>,
+    options: ResolverOpts,
+    on_send: O,
+) -> MockedNameServerPool<O> {
+    let conn_provider = MockConnProvider {
+        on_send: on_send.clone(),
+    };
+
     #[cfg(not(feature = "mdns"))]
-    return NameServerPool::from_nameservers(&options, udp, tcp);
+    return NameServerPool::from_nameservers(&options, udp, tcp, conn_provider);
 
     #[cfg(feature = "mdns")]
     return NameServerPool::from_nameservers(
         &options,
         udp,
         tcp,
-        _mdns.unwrap_or_else(|| mock_nameserver(vec![], options)),
+        _mdns.unwrap_or_else(|| mock_nameserver_on_send(vec![], options, on_send)),
+        conn_provider,
     );
 }
 
@@ -287,4 +332,213 @@ fn test_distrust_nx_responses() {
 
     let response = reactor.block_on(future).unwrap();
     assert_eq!(response.answers()[0], v4_record);
+}
+
+// === Concurrent requests ===
+
+#[derive(Clone)]
+struct OnSendBarrier {
+    barrier: Arc<AtomicIsize>,
+}
+
+impl OnSendBarrier {
+    fn new(count: isize) -> Self {
+        Self {
+            barrier: Arc::new(AtomicIsize::new(count)),
+        }
+    }
+}
+
+impl OnSend for OnSendBarrier {
+    fn on_send(
+        &mut self,
+        response: Result<DnsResponse, ProtoError>,
+    ) -> Box<Future<Item = DnsResponse, Error = ProtoError> + Send> {
+        self.barrier.fetch_sub(1, Ordering::Relaxed);
+
+        // loop until the barrier is 0
+        let loop_future = future::loop_fn(
+            (self.barrier.clone(), response),
+            move |(barrier, response)| {
+                let remaining = barrier.load(Ordering::Relaxed);
+
+                match remaining {
+                    0 => response.map(Loop::Break),
+                    i if i > 0 => Ok(Loop::Continue((barrier, response))),
+                    i if i < 0 => panic!("more concurrency than expected: {}", i),
+                    _ => panic!("all other cases handled"),
+                }
+            },
+        );
+
+        Box::new(loop_future)
+    }
+}
+
+#[test]
+fn test_concurrent_requests() {
+    let mut options = ResolverOpts::default();
+    // there are only 2 conns, so this matches that count
+    options.num_concurrent_reqs = 2;
+
+    // we want to make sure that both udp connections are called
+    //   this will count down to 0 only if both are called.
+    let on_send = OnSendBarrier::new(2);
+
+    let query = Query::query(Name::from_str("www.example.com.").unwrap(), RecordType::A);
+
+    let udp_record = v4_record(query.name().clone(), Ipv4Addr::new(127, 0, 0, 1));
+
+    let udp_message = message(query.clone(), vec![udp_record.clone()], vec![], vec![]);
+
+    let mut reactor = Runtime::new().unwrap();
+
+    let udp1_nameserver =
+        mock_nameserver_on_send(vec![udp_message.map(Into::into)], options, on_send.clone());
+    let udp2_nameserver = udp1_nameserver.clone();
+
+    let mut pool = mock_nameserver_pool_on_send(
+        vec![udp2_nameserver, udp1_nameserver],
+        vec![],
+        None,
+        options,
+        on_send,
+    );
+
+    // lookup on UDP succeeds, any other would fail
+    let request = message(query, vec![], vec![], vec![]).unwrap();
+    let future = pool.send(request);
+
+    // there's no actual network traffic happening, 1 sec should be plenty
+    //   TODO: for some reason this timout doesn't work, not clear why...
+    // let future = Timeout::new(future, Duration::from_secs(1));
+
+    let response = reactor.block_on(future).unwrap();
+    assert_eq!(response.answers()[0], udp_record);
+}
+
+#[test]
+fn test_concurrent_requests_more_than_conns() {
+    let mut options = ResolverOpts::default();
+    // there are only two conns, but this requests 3 concurrent requests, only 2 called
+    options.num_concurrent_reqs = 3;
+
+    // we want to make sure that both udp connections are called
+    //   this will count down to 0 only if both are called.
+    let on_send = OnSendBarrier::new(2);
+
+    let query = Query::query(Name::from_str("www.example.com.").unwrap(), RecordType::A);
+
+    let udp_record = v4_record(query.name().clone(), Ipv4Addr::new(127, 0, 0, 1));
+
+    let udp_message = message(query.clone(), vec![udp_record.clone()], vec![], vec![]);
+
+    let mut reactor = Runtime::new().unwrap();
+
+    let udp1_nameserver =
+        mock_nameserver_on_send(vec![udp_message.map(Into::into)], options, on_send.clone());
+    let udp2_nameserver = udp1_nameserver.clone();
+
+    let mut pool = mock_nameserver_pool_on_send(
+        vec![udp2_nameserver, udp1_nameserver],
+        vec![],
+        None,
+        options,
+        on_send,
+    );
+
+    // lookup on UDP succeeds, any other would fail
+    let request = message(query, vec![], vec![], vec![]).unwrap();
+    let future = pool.send(request);
+
+    // there's no actual network traffic happening, 1 sec should be plenty
+    //   TODO: for some reason this timout doesn't work, not clear why...
+    // let future = Timeout::new(future, Duration::from_secs(1));
+
+    let response = reactor.block_on(future).unwrap();
+    assert_eq!(response.answers()[0], udp_record);
+}
+
+#[test]
+fn test_concurrent_requests_1_conn() {
+    let mut options = ResolverOpts::default();
+    // there are two connections, but no concurrency requested
+    options.num_concurrent_reqs = 1;
+
+    // we want to make sure that both udp connections are called
+    //   this will count down to 0 only if both are called.
+    let on_send = OnSendBarrier::new(1);
+
+    let query = Query::query(Name::from_str("www.example.com.").unwrap(), RecordType::A);
+
+    let udp_record = v4_record(query.name().clone(), Ipv4Addr::new(127, 0, 0, 1));
+
+    let udp_message = message(query.clone(), vec![udp_record.clone()], vec![], vec![]);
+
+    let mut reactor = Runtime::new().unwrap();
+
+    let udp1_nameserver =
+        mock_nameserver_on_send(vec![udp_message.map(Into::into)], options, on_send.clone());
+    let udp2_nameserver = udp1_nameserver.clone();
+
+    let mut pool = mock_nameserver_pool_on_send(
+        vec![udp2_nameserver, udp1_nameserver],
+        vec![],
+        None,
+        options,
+        on_send,
+    );
+
+    // lookup on UDP succeeds, any other would fail
+    let request = message(query, vec![], vec![], vec![]).unwrap();
+    let future = pool.send(request);
+
+    // there's no actual network traffic happening, 1 sec should be plenty
+    //   TODO: for some reason this timout doesn't work, not clear why...
+    // let future = Timeout::new(future, Duration::from_secs(1));
+
+    let response = reactor.block_on(future).unwrap();
+    assert_eq!(response.answers()[0], udp_record);
+}
+
+#[test]
+fn test_concurrent_requests_0_conn() {
+    let mut options = ResolverOpts::default();
+    // there are two connections, but no concurrency requested, 0==1
+    options.num_concurrent_reqs = 0;
+
+    // we want to make sure that both udp connections are called
+    //   this will count down to 0 only if both are called.
+    let on_send = OnSendBarrier::new(1);
+
+    let query = Query::query(Name::from_str("www.example.com.").unwrap(), RecordType::A);
+
+    let udp_record = v4_record(query.name().clone(), Ipv4Addr::new(127, 0, 0, 1));
+
+    let udp_message = message(query.clone(), vec![udp_record.clone()], vec![], vec![]);
+
+    let mut reactor = Runtime::new().unwrap();
+
+    let udp1_nameserver =
+        mock_nameserver_on_send(vec![udp_message.map(Into::into)], options, on_send.clone());
+    let udp2_nameserver = udp1_nameserver.clone();
+
+    let mut pool = mock_nameserver_pool_on_send(
+        vec![udp2_nameserver, udp1_nameserver],
+        vec![],
+        None,
+        options,
+        on_send,
+    );
+
+    // lookup on UDP succeeds, any other would fail
+    let request = message(query, vec![], vec![], vec![]).unwrap();
+    let future = pool.send(request);
+
+    // there's no actual network traffic happening, 1 sec should be plenty
+    //   TODO: for some reason this timout doesn't work, not clear why...
+    // let future = Timeout::new(future, Duration::from_secs(1));
+
+    let response = reactor.block_on(future).unwrap();
+    assert_eq!(response.answers()[0], udp_record);
 }