+++ crates/resolver/CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.11
 
+### Added
+
+- New option to execute queries concurrently, default is 2 #615
+
 ### Changed
 
 - Added option to distrust Nameservers on SERVFAIL responses, continue resolution #613